+++ CHANGELOG.md
@@ -178,3 +178,4 @@
 - feat: remove `key` from message struct, the `key` info should be migrate to `params` properties
 - feat: add document for english and chinese language
 - feat: add default workflow version
+- fix: fix issue #16 not save the data after setting the process vars